update: 更新文件 manage_auto_start_apps_serializer.cpp 定义变量类型修改

+++ services/edm_plugin/src/utils/manage_auto_start_apps_serializer.cpp
@@ -74,7 +74,7 @@ std::vector<ManageAutoStartAppInfo> ManageAutoStartAppsSerializer::SetIntersecti
     std::vector<ManageAutoStartAppInfo> mergeData;
     for (const std::string &dataItem : data) {
         for (const ManageAutoStartAppInfo &item : currentData) {
-            int32_t index = dataItem.rfind("/");
+            size_t index = dataItem.rfind("/");
             std::string uniqueKey = dataItem.substr(0, index);
             if (item.GetUniqueKey() == uniqueKey) {
                 mergeData.push_back(item);
